move PublicHandle from main.rs to clashlib
